A new boundary condition has been added: CO2 map for all the Med Sea.
At the moment we use an uniform map instead of a single value - the result is the same.
But we prepare the model to accept a 2 dimensional field

Now, the time interpolation is done for the co2 as well as is done for others boundary conditions 
such as riv, gib etc. 
Then, time and values are not hardcoded !! (as was before for co2)

These changes have became urgent by modifications in operative chain.