adding a new flag in namelist.diagnostics, called save_bkp_group2
If false, ave.backups of group 2 (low frequencies) will not be saved.

In theory, there is no reason to set is as .false., 
because ave of group 2 would not take in account averaging before restart.
BUT, in the specific case of operational chain V4, 
an only restart time at 12:00 (of Friday, for example), 
can cause the dump of 12 GB of unuseful files.

That's why we prefer that the user set namelist values.