IMPORTANT BUG FIX: update every time step of all time dependent variables. At the moment some variables depending on time are not updated every time step, and maintain the value at time of initialization. These are
zbtr_arr
inv_eu
inv_ev
inv_et

for trcadv, and

zbtr
zeeu
zeev
for trchdf.

This is obviously a bug, and results the fact that the final result of a ogstm simulation depends on interruptions/restarts we do along the simulation.

Actually,
- zbtr_arr
- inv_eu
- inv_ev
- inv_et
were calculated as 3D arrays in the initialization of trcadv
and only the surface was updated every time step.
Now we do the 3D loop every time step.

There is still an unresolved bug.
Now the final result of a simulation does not depend on the number of restarts only if we edit trcnxt.f90 with this comment:

-            tra(jk,jj,ji,jn  ) = tra(jk,jj,ji,jn  )*e3t_back(jk,jj,ji)/e3t(jk,jj,ji)
+            tra(jk,jj,ji,jn  ) = tra(jk,jj,ji,jn  )! *e3t_back(jk,jj,ji)/e3t(jk,jj,ji)

Because e3t_back is not saved when we save restarts.
This bug will be soon fixed (hopefully!)