Implementation of variable deltaT
Delta_T files are provided in wrkdir/MODEL/ as
DELTA_T/DeltaT_"//datestring//".txt

In forcing_phys.f90
DeltaT file is loaded in load_phys()
and an integer imposed_deltaT (possible values 1800,900,600,450,360,300) is read.

We use the same approach of all the other forcings, there are before and after values
So we have:
imposed_deltaT (before) and imposed_deltaT (after)
As rdt we choose the minimum, with a type cast to have a real*8.

IMPORTANT DETAIL
To be sure to have the change of rdt at 01:00:00, 02:00:00 (not 12:07:30)
TimeManager.TimeInterpolation() has been modified in WHILE loop

STRUCT%Times(I) < sec —> STRUCT%Times(I) <= sec

For example, for that simulation time: 20180101 —> 20180107
Behavior with “<“:
- at the first timestep ogstm.xx searches for 20171231
- the last timestep ogstm.xx searches for 20180107

Behavior with “<=“:
- at the first timestep ogstm.xx searches for 20180101
- the last timestep ogstm searches for 20180108

So, now we need to have a link of the last day.