I/O coherence
In both diadump.f90 and trcdit.f90 counters are erased from cycles of dumping: the variable to store is defined from matrices of indexes and counters.
Proper variable to dump is controlled inside each loop from selecting the correct row and column of the matrix.
Cleaning of code from old print and old counters.